regrouped all fluidstack useful methods to the same class and moved it api
